+++ src/apple/apple-wizard.ts
@@ -7,7 +7,7 @@
 import clack from '@clack/prompts';
 import * as fs from 'fs';
 import * as path from 'path';
-import * as xcManager from './xcode-manager';
+import { XcodeProject } from './xcode-manager';
 import * as codeTools from './code-tools';
 import * as bash from '../utils/bash';
 import * as SentryUtils from '../utils/sentrycli-utils';
@@ -25,6 +25,7 @@ import {
   printWelcome,
   abort,
   askForItemSelection,
+  confirmContinueEvenThoughNoGitRepo,
   getOrAskForProjectData,
 } from '../utils/clack-utils';
 
@@ -46,6 +47,8 @@ async function runAppleWizardWithTelementry(
     promoCode: options.promoCode,
   });
 
+  await confirmContinueEvenThoughNoGitRepo();
+
   const hasCli = bash.hasSentryCLI();
   Sentry.setTag('has-cli', hasCli);
   if (!hasCli) {
@@ -63,7 +66,7 @@ async function runAppleWizardWithTelementry(
   }
 
   const projectDir = process.cwd();
-  const xcodeProjFiles = findFilesWithExtension(projectDir, '.xcodeproj');
+  const xcodeProjFiles = searchXcodeProject(projectDir);
 
   if (!xcodeProjFiles || xcodeProjFiles.length === 0) {
     clack.log.error(
@@ -91,13 +94,36 @@ async function runAppleWizardWithTelementry(
   }
 
   const pbxproj = path.join(projectDir, xcodeProjFile, 'project.pbxproj');
+
   if (!fs.existsSync(pbxproj)) {
-    clack.log.error(`No pbxproj found at ${pbxproj}`);
+    clack.log.error(`No pbxproj found at ${xcodeProjFile}`);
     await abort();
     return;
   }
 
   const { project, apiKey } = await getSentryProjectAndApiKey(options);
+
+  const xcProject = new XcodeProject(pbxproj);
+
+  const availableTargets = xcProject.getAllTargets();
+
+  if (availableTargets.length == 0) {
+    clack.log.error(`No suttable target found in ${xcodeProjFile}`);
+    await abort();
+    return;
+  }
+
+  const target =
+    availableTargets.length == 1
+      ? availableTargets[0]
+      : (
+          await traceStep('Choose target', () =>
+            askForItemSelection(
+              availableTargets,
+              'Which target do you want to add Sentry to?',
+            ),
+          )
+        ).value;
 
   SentryUtils.createSentryCLIRC(projectDir, { auth_token: apiKey.token });
   clack.log.info(
@@ -130,17 +156,17 @@ async function runAppleWizardWithTelementry(
   }
 
   traceStep('Update Xcode project', () => {
-    xcManager.updateXcodeProject(pbxproj, project, !hasCocoa, true);
+    xcProject.updateXcodeProject(project, target, apiKey, !hasCocoa, true);
   });
 
   Sentry.setTag('package-manager', hasCocoa ? 'cocoapods' : 'SPM');
-  const projSource = path.join(
-    projectDir,
-    xcodeProjFile.replace('.xcodeproj', ''),
-  );
   const codeAdded = traceStep('Add code snippet', () => {
+    const files = xcProject.filesForTarget(target);
+    if (files === undefined || files.length == 0) return false;
+
     return codeTools.addCodeSnippetToProject(
-      projSource,
+      projectDir,
+      files,
       project.keys[0].dsn.public,
     );
   });
@@ -157,17 +183,28 @@ async function runAppleWizardWithTelementry(
         'Found a Fastfile in your project. Do you want to configure a lane to upload debug symbols to Sentry?',
     });
     if (addLane) {
-      await traceStep('Configure fastlane', () =>
+      const added = await traceStep('Configure fastlane', () =>
         fastlane.addSentryToFastlane(
           projectDir,
           project.organization.slug,
           project.slug,
         ),
       );
+      if (added) {
+        clack.log.step(
+          'A new step was added to your fastlane file. Now and you build your project with fastlane, debug symbols and source context will be uploaded to Sentry.',
+        );
+      } else {
+        clack.log.warn(
+          'Could not edit your fastlane file to upload debug symbols to Sentry. Please follow the instructions at https://docs.sentry.io/platforms/apple/guides/ios/dsym/#fastlane',
+        );
+      }
     }
   }
 
-  clack.log.success('Sentry was successfully added to your project!');
+  clack.log.success(
+    'Sentry was successfully added to your project! Run your project to send your first event to Sentry. Go to Sentry.io to see whether everything is working fine.',
+  );
 }
 
 //Prompt for Sentry project and API key
@@ -176,6 +213,44 @@ async function getSentryProjectAndApiKey(
 ): Promise<{ project: SentryProjectData; apiKey: { token: string } }> {
   const { selectedProject, authToken } = await getOrAskForProjectData(options);
   return { project: selectedProject, apiKey: { token: authToken } };
+}
+
+function searchXcodeProject(at: string): string[] {
+  const projs = findFilesWithExtension(at, '.xcodeproj');
+  if (projs.length > 0) {
+    return projs;
+  }
+
+  const workspace = findFilesWithExtension(at, '.xcworkspace');
+  if (workspace.length == 0) {
+    return [];
+  }
+
+  const xsworkspacedata = path.join(
+    at,
+    workspace[0],
+    'contents.xcworkspacedata',
+  );
+  if (!fs.existsSync(xsworkspacedata)) {
+    return [];
+  }
+  const groupRegex = /location *= *"group:([^"]+)"/gim;
+  const content = fs.readFileSync(xsworkspacedata, 'utf8');
+  let matches = groupRegex.exec(content);
+
+  while (matches) {
+    const group = matches[1];
+    const groupPath = path.join(at, group);
+    if (
+      !group.endsWith('Pods.xcodeproj') &&
+      group.endsWith('.xcodeproj') &&
+      fs.existsSync(groupPath)
+    ) {
+      projs.push(group);
+    }
+    matches = groupRegex.exec(content);
+  }
+  return projs;
 }
 
 //find files with the given extension